+++ google-cloud-beyond_corp-app_connections-v1/lib/google/cloud/beyond_corp/app_connections/v1/app_connections_service.rb
@@ -33,12 +33,12 @@ module Google
       module AppConnections
         module V1
           ##
-          # ## API Overview
+          # API Overview:
           #
           # The `beyondcorp.googleapis.com` service implements the Google Cloud
           # BeyondCorp API.
           #
-          # ## Data Model
+          # Data Model:
           #
           # The AppConnectionsService exposes the following resources:
           #