+++ google-cloud-beyond_corp/lib/google/cloud/beyond_corp/app_connectors.rb
@@ -58,12 +58,12 @@ module Google
         #
         # ## About AppConnectorsService
         #
-        # ## API Overview
+        # API Overview:
         #
         # The `beyondcorp.googleapis.com` service implements the Google Cloud
         # BeyondCorp API.
         #
-        # ## Data Model
+        # Data Model:
         #
         # The AppConnectorsService exposes the following resource:
         #