+++ google-cloud-bigquery-data_transfer-v1/proto_docs/google/cloud/bigquery/datatransfer/v1/transfer.rb
@@ -88,7 +88,9 @@ module Google
           #     User specified display name for the data transfer.
           # @!attribute [rw] data_source_id
           #   @return [::String]
-          #     Data source id. Cannot be changed once data transfer is created.
+          #     Data source ID. This cannot be changed once data transfer is created. The
+          #     full list of available data source IDs can be returned through an API call:
+          #     https://cloud.google.com/bigquery-transfer/docs/reference/datatransfer/rest/v1/projects.locations.dataSources/list
           # @!attribute [rw] params
           #   @return [::Google::Protobuf::Struct]
           #     Parameters specific to each data source. For more information see the