+++ google-cloud-datastore/acceptance/datastore/datastore_test.rb
@@ -815,4 +815,58 @@ describe Google::Cloud::Datastore::Dataset, :datastore do
       _(refresh).must_be :nil?
     end
   end
+
+  describe "querying with limit > 300" do
+    let(:kind_val) { "Post_#{SecureRandom.hex(4)}" }
+    let(:limit) { 700 }
+    let(:post) do
+      Google::Cloud::Datastore::Entity.new.tap do |e|
+        e["title"]       = "How to make the perfect pizza in your grill"
+      end
+    end
+
+    before :all do
+      # Add 1000 entities of the same kind to the datastore
+      1000.times.each do |id|
+        post_temp = post.dup
+        post_temp.key = Google::Cloud::Datastore::Key.new kind_val, "Post_#{id+1}"
+        dataset.save post_temp
+      end
+    end
+
+    after :all do
+      # Delete the entities added
+      1000.times.each do |id|
+        post_temp = post.dup
+        post_temp.key = Google::Cloud::Datastore::Key.new kind_val, "Post_#{id+1}"
+        dataset.delete post_temp
+      end
+    end
+
+    it "should limit results when limit > 300 in query" do
+      # Testing limit with query
+      query = dataset.query(kind_val).limit(limit)
+      entities_count = 0
+      results = dataset.run query
+      loop do
+        entities_count += results.count
+        break unless results.next?
+        results = results.next
+      end
+      _(entities_count).must_equal limit
+    end
+
+    it "should limit results when limit > 300 in GQL query" do
+      # Testing limit with GQL query
+      query_gql = dataset.gql "SELECT * FROM #{kind_val} LIMIT @limit", {limit: limit}
+      entities_count = 0
+      results = dataset.run query_gql
+      loop do
+        entities_count += results.count
+        break unless results.next?
+        results = results.next
+      end
+      _(entities_count).must_equal limit
+    end
+  end
 end