+++ google-cloud-datastore/acceptance/datastore_helper.rb
@@ -20,6 +20,7 @@ require "minitest/focus"
 require "minitest/rg"
 require "google/cloud/datastore"
 require "securerandom"
+require "minitest/hooks/default"
 
 # Create shared dataset object so we don't create new for each test
 $dataset = Google::Cloud.new.datastore