+++ google-cloud-dialogflow-cx-v3/proto_docs/google/cloud/dialogflow/cx/v3/security_settings.rb
@@ -164,12 +164,11 @@ module Google
           #     `SecuritySettings`.
           # @!attribute [rw] retention_window_days
           #   @return [::Integer]
-          #     Retains data in interaction logging for the specified number of days.
-          #     This does not apply to Cloud logging, which is owned by the user - not
-          #     Dialogflow.
-          #     User must set a value lower than Dialogflow's default 365d TTL (time to
-          #     live). Setting a value higher than that has no effect. A missing value or
-          #     setting to 0 also means we use Dialogflow's default TTL.
+          #     Retains the data for the specified number of days.
+          #     User must set a value lower than Dialogflow's default 365d TTL (30 days
+          #     for Agent Assist traffic), higher value will be ignored and use default.
+          #     Setting a value higher than that has no effect. A missing value or
+          #     setting to 0 also means we use default TTL.
           # @!attribute [rw] purge_data_types
           #   @return [::Array<::Google::Cloud::Dialogflow::CX::V3::SecuritySettings::PurgeDataType>]
           #     List of types of data to remove when retention settings triggers purge.