+++ google-cloud-memcache-v1beta2/lib/google/cloud/memcache/v1beta2/cloud_memcache/client.rb
@@ -18,6 +18,7 @@
 
 require "google/cloud/errors"
 require "google/cloud/memcache/v1beta2/cloud_memcache_pb"
+require "google/cloud/location"
 
 module Google
   module Cloud
@@ -94,6 +95,8 @@ module Google
 
                 default_config.rpcs.apply_software_update.timeout = 1200.0
 
+                default_config.rpcs.reschedule_maintenance.timeout = 1200.0
+
                 default_config
               end
               yield @configure if block_given?
@@ -169,6 +172,12 @@ module Google
                 config.endpoint = @config.endpoint
               end
 
+              @location_client = Google::Cloud::Location::Locations::Client.new do |config|
+                config.credentials = credentials
+                config.quota_project = @quota_project_id
+                config.endpoint = @config.endpoint
+              end
+
               @cloud_memcache_stub = ::Gapic::ServiceStub.new(
                 ::Google::Cloud::Memcache::V1beta2::CloudMemcache::Stub,
                 credentials:  credentials,
@@ -184,6 +193,13 @@ module Google
             # @return [::Google::Cloud::Memcache::V1beta2::CloudMemcache::Operations]
             #
             attr_reader :operations_client
+
+            ##
+            # Get the associated client for mix-in of the Locations.
+            #
+            # @return [Google::Cloud::Location::Locations::Client]
+            #
+            attr_reader :location_client
 
             # Service calls
 
@@ -511,6 +527,7 @@ module Google
             #
             #   @param update_mask [::Google::Protobuf::FieldMask, ::Hash]
             #     Required. Mask of fields to update.
+            #
             #      *  `displayName`
             #   @param resource [::Google::Cloud::Memcache::V1beta2::Instance, ::Hash]
             #     Required. A Memcached [Instance] resource.
@@ -987,6 +1004,107 @@ module Google
             end
 
             ##
+            # Performs the apply phase of the RescheduleMaintenance verb.
+            #
+            # @overload reschedule_maintenance(request, options = nil)
+            #   Pass arguments to `reschedule_maintenance` via a request object, either of type
+            #   {::Google::Cloud::Memcache::V1beta2::RescheduleMaintenanceRequest} or an equivalent Hash.
+            #
+            #   @param request [::Google::Cloud::Memcache::V1beta2::RescheduleMaintenanceRequest, ::Hash]
+            #     A request object representing the call parameters. Required. To specify no
+            #     parameters, or to keep all the default parameter values, pass an empty Hash.
+            #   @param options [::Gapic::CallOptions, ::Hash]
+            #     Overrides the default settings for this call, e.g, timeout, retries, etc. Optional.
+            #
+            # @overload reschedule_maintenance(instance: nil, reschedule_type: nil, schedule_time: nil)
+            #   Pass arguments to `reschedule_maintenance` via keyword arguments. Note that at
+            #   least one keyword argument is required. To specify no parameters, or to keep all
+            #   the default parameter values, pass an empty Hash as a request object (see above).
+            #
+            #   @param instance [::String]
+            #     Required. Memcache instance resource name using the form:
+            #         `projects/{project_id}/locations/{location_id}/instances/{instance_id}`
+            #     where `location_id` refers to a GCP region.
+            #   @param reschedule_type [::Google::Cloud::Memcache::V1beta2::RescheduleMaintenanceRequest::RescheduleType]
+            #     Required. If reschedule type is SPECIFIC_TIME, must set up schedule_time as well.
+            #   @param schedule_time [::Google::Protobuf::Timestamp, ::Hash]
+            #     Timestamp when the maintenance shall be rescheduled to if
+            #     reschedule_type=SPECIFIC_TIME, in RFC 3339 format, for
+            #     example `2012-11-15T16:19:00.094Z`.
+            #
+            # @yield [response, operation] Access the result along with the RPC operation
+            # @yieldparam response [::Gapic::Operation]
+            # @yieldparam operation [::GRPC::ActiveCall::Operation]
+            #
+            # @return [::Gapic::Operation]
+            #
+            # @raise [::Google::Cloud::Error] if the RPC is aborted.
+            #
+            # @example Basic example
+            #   require "google/cloud/memcache/v1beta2"
+            #
+            #   # Create a client object. The client can be reused for multiple calls.
+            #   client = Google::Cloud::Memcache::V1beta2::CloudMemcache::Client.new
+            #
+            #   # Create a request. To set request fields, pass in keyword arguments.
+            #   request = Google::Cloud::Memcache::V1beta2::RescheduleMaintenanceRequest.new
+            #
+            #   # Call the reschedule_maintenance method.
+            #   result = client.reschedule_maintenance request
+            #
+            #   # The returned object is of type Gapic::Operation. You can use this
+            #   # object to check the status of an operation, cancel it, or wait
+            #   # for results. Here is how to block until completion:
+            #   result.wait_until_done! timeout: 60
+            #   if result.response?
+            #     p result.response
+            #   else
+            #     puts "Error!"
+            #   end
+            #
+            def reschedule_maintenance request, options = nil
+              raise ::ArgumentError, "request must be provided" if request.nil?
+
+              request = ::Gapic::Protobuf.coerce request, to: ::Google::Cloud::Memcache::V1beta2::RescheduleMaintenanceRequest
+
+              # Converts hash and nil to an options object
+              options = ::Gapic::CallOptions.new(**options.to_h) if options.respond_to? :to_h
+
+              # Customize the options with defaults
+              metadata = @config.rpcs.reschedule_maintenance.metadata.to_h
+
+              # Set x-goog-api-client and x-goog-user-project headers
+              metadata[:"x-goog-api-client"] ||= ::Gapic::Headers.x_goog_api_client \
+                lib_name: @config.lib_name, lib_version: @config.lib_version,
+                gapic_version: ::Google::Cloud::Memcache::V1beta2::VERSION
+              metadata[:"x-goog-user-project"] = @quota_project_id if @quota_project_id
+
+              header_params = {}
+              if request.instance
+                header_params["instance"] = request.instance
+              end
+
+              request_params_header = header_params.map { |k, v| "#{k}=#{v}" }.join("&")
+              metadata[:"x-goog-request-params"] ||= request_params_header
+
+              options.apply_defaults timeout:      @config.rpcs.reschedule_maintenance.timeout,
+                                     metadata:     metadata,
+                                     retry_policy: @config.rpcs.reschedule_maintenance.retry_policy
+
+              options.apply_defaults timeout:      @config.timeout,
+                                     metadata:     @config.metadata,
+                                     retry_policy: @config.retry_policy
+
+              @cloud_memcache_stub.call_rpc :reschedule_maintenance, request, options: options do |response, operation|
+                response = ::Gapic::Operation.new response, @operations_client, options: options
+                yield response, operation if block_given?
+                return response
+              end
+            rescue ::GRPC::BadStatus => e
+              raise ::Google::Cloud::Error.from_error(e)
+            end
+
+            ##
             # Configuration class for the CloudMemcache API.
             #
             # This class represents the configuration for CloudMemcache,
@@ -1161,6 +1279,11 @@ module Google
                 # @return [::Gapic::Config::Method]
                 #
                 attr_reader :apply_software_update
+                ##
+                # RPC-specific configuration for `reschedule_maintenance`
+                # @return [::Gapic::Config::Method]
+                #
+                attr_reader :reschedule_maintenance
 
                 # @private
                 def initialize parent_rpcs = nil
@@ -1180,6 +1303,8 @@ module Google
                   @apply_parameters = ::Gapic::Config::Method.new apply_parameters_config
                   apply_software_update_config = parent_rpcs.apply_software_update if parent_rpcs.respond_to? :apply_software_update
                   @apply_software_update = ::Gapic::Config::Method.new apply_software_update_config
+                  reschedule_maintenance_config = parent_rpcs.reschedule_maintenance if parent_rpcs.respond_to? :reschedule_maintenance
+                  @reschedule_maintenance = ::Gapic::Config::Method.new reschedule_maintenance_config
 
                   yield self if block_given?
                 end