+++ google-cloud-os_login-v1/test/google/cloud/os_login/v1/os_login_service_test.rb
@@ -47,6 +47,66 @@ class ::Google::Cloud::OsLogin::V1::OsLoginService::ClientTest < Minitest::Test
     end
   end
 
+  def test_create_ssh_public_key
+    # Create GRPC objects.
+    grpc_response = ::Google::Cloud::OsLogin::Common::SshPublicKey.new
+    grpc_operation = GRPC::ActiveCall::Operation.new nil
+    grpc_channel = GRPC::Core::Channel.new "localhost:8888", nil, :this_channel_is_insecure
+    grpc_options = {}
+
+    # Create request parameters for a unary method.
+    parent = "hello world"
+    ssh_public_key = {}
+
+    create_ssh_public_key_client_stub = ClientStub.new grpc_response, grpc_operation do |name, request, options:|
+      assert_equal :create_ssh_public_key, name
+      assert_kind_of ::Google::Cloud::OsLogin::V1::CreateSshPublicKeyRequest, request
+      assert_equal "hello world", request["parent"]
+      assert_equal Gapic::Protobuf.coerce({}, to: ::Google::Cloud::OsLogin::Common::SshPublicKey), request["ssh_public_key"]
+      refute_nil options
+    end
+
+    Gapic::ServiceStub.stub :new, create_ssh_public_key_client_stub do
+      # Create client
+      client = ::Google::Cloud::OsLogin::V1::OsLoginService::Client.new do |config|
+        config.credentials = grpc_channel
+      end
+
+      # Use hash object
+      client.create_ssh_public_key({ parent: parent, ssh_public_key: ssh_public_key }) do |response, operation|
+        assert_equal grpc_response, response
+        assert_equal grpc_operation, operation
+      end
+
+      # Use named arguments
+      client.create_ssh_public_key parent: parent, ssh_public_key: ssh_public_key do |response, operation|
+        assert_equal grpc_response, response
+        assert_equal grpc_operation, operation
+      end
+
+      # Use protobuf object
+      client.create_ssh_public_key ::Google::Cloud::OsLogin::V1::CreateSshPublicKeyRequest.new(parent: parent, ssh_public_key: ssh_public_key) do |response, operation|
+        assert_equal grpc_response, response
+        assert_equal grpc_operation, operation
+      end
+
+      # Use hash object with options
+      client.create_ssh_public_key({ parent: parent, ssh_public_key: ssh_public_key }, grpc_options) do |response, operation|
+        assert_equal grpc_response, response
+        assert_equal grpc_operation, operation
+      end
+
+      # Use protobuf object with options
+      client.create_ssh_public_key(::Google::Cloud::OsLogin::V1::CreateSshPublicKeyRequest.new(parent: parent, ssh_public_key: ssh_public_key), grpc_options) do |response, operation|
+        assert_equal grpc_response, response
+        assert_equal grpc_operation, operation
+      end
+
+      # Verify method calls
+      assert_equal 5, create_ssh_public_key_client_stub.call_rpc_count
+    end
+  end
+
   def test_delete_posix_account
     # Create GRPC objects.
     grpc_response = ::Google::Protobuf::Empty.new