+++ google-cloud-service_management-v1/proto_docs/google/api/quota.rb
@@ -41,10 +41,10 @@ module Google
     #          STANDARD: 10000
     #
     #
-    #      # The metric rules bind all methods to the read_calls metric,
-    #      # except for the UpdateBook and DeleteBook methods. These two methods
-    #      # are mapped to the write_calls metric, with the UpdateBook method
-    #      # consuming at twice rate as the DeleteBook method.
+    #      (The metric rules bind all methods to the read_calls metric,
+    #       except for the UpdateBook and DeleteBook methods. These two methods
+    #       are mapped to the write_calls metric, with the UpdateBook method
+    #       consuming at twice rate as the DeleteBook method.)
     #      metric_rules:
     #      - selector: "*"
     #        metric_costs:
@@ -70,10 +70,10 @@ module Google
     #        value_type: INT64
     # @!attribute [rw] limits
     #   @return [::Array<::Google::Api::QuotaLimit>]
-    #     List of `QuotaLimit` definitions for the service.
+    #     List of QuotaLimit definitions for the service.
     # @!attribute [rw] metric_rules
     #   @return [::Array<::Google::Api::MetricRule>]
-    #     List of `MetricRule` definitions, each one mapping a selected method to one
+    #     List of MetricRule definitions, each one mapping a selected method to one
     #     or more metrics.
     class Quota
       include ::Google::Protobuf::MessageExts